\name{ols}
\alias{ols}
\title{Ordinary Least Squares Regression}
\usage{
  ols(reg, preds = NULL)
}
\arguments{
  \item{\code{Regression}}{object to confirm}

  \item{\code{vector}}{the columns of the matrix to
  consider for regression}
}
\value{
  \code{list} of regression diagnostics
}
\description{
  Ordinary Least Squares Regression
}

