\name{regression}
\alias{regression}
\title{Linear Regression Object initializer}
\usage{
  \method{Regression}{initialize} (output = NULL,
    predictors = NULL)
}
\arguments{
  \item{output}{\code{numeric vector} of outputs to predict
  with regression}

  \item{predictors}{\code{numeric matrix} of inputs to use
  for prediction}
}
\value{
  \code{OLS} Regression object
}
\description{
  Linear Regression Object initializer
}

